Revise Font and FontFace structures
